Correcting the type of double for BigDecimal.

+++ src/main/java/com/ms/service/dto/ServiceDTO.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.beans.BeanUtils;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -25,7 +26,7 @@ public class ServiceDTO implements Serializable {
 
     private boolean serviceStatus; // status
     private String serviceDescription; // descrição
-    private double servicePrice; // preço
+    private BigDecimal servicePrice; // preço
     private int serviceRuntime; // tempo de execução
     private int serviceTerm; // prazo
     private String created;
